requirements.txt
fastapi
uvicorn[standard]
python-dotenv
solana==0.30.2       # solana-py
solders>=0.18.0,<0.19.0      # Solana data structures
requests
python-multipart     # file uploads for IPFS
PyJWT                #  JWT if we  add login
pynacl               #  signature verify for wallet sign-in
